Thêm BaseResponse

+++ shopapp/src/main/java/com/example/shopapp/service/impl/ICategoryService.java
@@ -1,34 +1,37 @@
 package com.example.shopapp.service.impl;
 
 import com.example.shopapp.dto.CategoryDTO;
+import com.example.shopapp.entity.BaseResponse;
 import com.example.shopapp.entity.Category;
 import com.example.shopapp.repository.CategoryRepository;
 import com.example.shopapp.service.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class ICategoryService implements CategoryService {
+public class ICategoryService extends BaseResponse<Category> implements CategoryService {
 
     private final CategoryRepository categoryRepository;
 
     private final ModelMapper mapper;
+
     @Override
-    public Category getCategoryById(Long categoryId) {
+    public BaseResponse<Category> getCategoryById(Long categoryId) {
+        BaseResponse<Category> response = new BaseResponse<>("Không tìm thấy danh mục nào có id = " + categoryId, null);
         Category category = categoryRepository.findById(categoryId).orElse(null);
-        return category;
+        if (category != null) {
+            response.setData(category);
+            response.setMessage("Đã tìm thấy danh mục " + category.getName());
+        }
+        return response;
     }
 
     @Override
     public List<Category> getAllCategories(int pageIndex, int pageSize) {
-//        Pageable pageable = new PageRequest(pageIndex, pageSize, new Sort());
         List<Category> listCategories = categoryRepository.findAll();
         return listCategories;
     }
@@ -42,7 +45,7 @@ public class ICategoryService implements CategoryService {
 
     @Override
     public Category updateCategory(CategoryDTO categoryDTO, Long id) {
-        Category categoryExist = getCategoryById(id);
+        Category categoryExist = categoryRepository.findById(id).orElse(null);
         if (categoryExist != null) {
             categoryExist.setName(categoryDTO.getName());
             categoryExist = categoryRepository.save(categoryExist);
@@ -54,8 +57,13 @@ public class ICategoryService implements CategoryService {
     public String deleteCategory(Long categoryId) {
         String message = "";
         try {
-            categoryRepository.deleteById(categoryId);
-            message = "Xóa danh mục thành công";
+            Category category = categoryRepository.findById(categoryId).orElse(null);
+            if (category != null) {
+                categoryRepository.deleteById(categoryId);
+                message = "Xóa danh mục thành công";
+            } else {
+                message = "Danh mục không tồn tại";
+            }
 
         } catch (Exception e) {
             message = e.getMessage();
